+++ corpan/packs/shared/capabilities/pronounce/contract.test.ts
@@ -1,0 +1,4 @@
+import { runContractSuite } from "../core/contractSuite"
+import { pronounceDriver } from "../test/drivers"
+
+runContractSuite(pronounceDriver)