+++ corpan/packs/shared/capabilities/segment-player/contract.test.ts
@@ -1,0 +1,4 @@
+import { runContractSuite } from "../core/contractSuite"
+import { segmentPlayerDriver } from "../test/drivers"
+
+runContractSuite(segmentPlayerDriver)